Move cqc ratings columns script
